// src/api/models/GetUserSellsShiftSuccessDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GetUserSellsShiftSuccessDto = {
    success: boolean;
    message: string;
    method: Record<string, any>;
    endpoint: string;
    statusCode: number;
    timestamp: string;
    data?: Array<string>;
};